backend/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-admin-shortcuts==2.0.0
django-ckeditor==5.6.1
django-colorfield==0.1.15
django-js-asset==1.1.0
gunicorn==19.9.0
idna==2.8
Pillow==5.4.1
psycopg2==2.7.7
pytz==2018.9
requests==2.21.0
selenium==3.141.0
singlemodeladmin==0.9
urllib3==1.24.1
